+++ core/tracing.go
@@ -29,7 +29,7 @@ func StartDutyTrace(ctx context.Context, duty Duty, spanName string, opts ...tra
 	ctx, outerSpan = tracer.Start(tracer.RootedCtx(ctx, traceID), fmt.Sprintf("core/duty.%s", strings.Title(duty.Type.String())))
 	ctx, innerSpan = tracer.Start(ctx, spanName, opts...)
 
-	outerSpan.SetAttributes(attribute.Int64("slot", duty.Slot))
+	outerSpan.SetAttributes(attribute.Int64("slot", int64(duty.Slot)))
 
 	return ctx, withEndSpan{
 		Span:    innerSpan,
@@ -77,13 +77,13 @@ func WithTracing() WireOption {
 
 			return clone.DutyDBStore(ctx, duty, set)
 		}
-		w.DutyDBAwaitAttestation = func(parent context.Context, slot, commIdx int64) (*eth2p0.AttestationData, error) {
+		w.DutyDBAwaitAttestation = func(parent context.Context, slot, commIdx uint64) (*eth2p0.AttestationData, error) {
 			ctx, span := tracer.Start(parent, "core/dutydb.AwaitAttestation")
 			defer span.End()
 
 			return clone.DutyDBAwaitAttestation(ctx, slot, commIdx)
 		}
-		w.DutyDBPubKeyByAttestation = func(parent context.Context, slot, commIdx, valCommIdx int64) (PubKey, error) {
+		w.DutyDBPubKeyByAttestation = func(parent context.Context, slot, commIdx, valCommIdx uint64) (PubKey, error) {
 			ctx, span := tracer.Start(parent, "core/dutydb.PubKeyByAttestation")
 			defer span.End()
